+++ src/icloudpd/authentication.py
@@ -6,6 +6,7 @@ from typing import Callable, Optional
 import click
 import pyicloud_ipd
 from pyicloud_ipd.base import PyiCloudService
+from pyicloud_ipd.raw_policy import RawTreatmentPolicy
 
 
 class TwoStepAuthRequiredError(Exception):
@@ -15,7 +16,7 @@ class TwoStepAuthRequiredError(Exception):
     """
 
 
-def authenticator(logger: logging.Logger, domain: str, lp_filename_generator: Callable[[str], str], use_raw_as_original:bool) -> Callable[[str, Optional[str], Optional[str], bool, Optional[str]], PyiCloudService]:
+def authenticator(logger: logging.Logger, domain: str, lp_filename_generator: Callable[[str], str], raw_policy:RawTreatmentPolicy) -> Callable[[str, Optional[str], Optional[str], bool, Optional[str]], PyiCloudService]:
     """Wraping authentication with domain context"""
     def authenticate_(
             username:str,
@@ -33,7 +34,7 @@ def authenticator(logger: logging.Logger, domain: str, lp_filename_generator: Ca
                 icloud = PyiCloudService(
                     lp_filename_generator,
                     domain,
-                    use_raw_as_original,
+                    raw_policy,
                     username, password,
                     cookie_directory=cookie_directory,
                     client_id=client_id,