+++ src/pyicloud_ipd/base.py
@@ -16,6 +16,7 @@ from pyicloud_ipd.exceptions import (
     PyiCloudAPIResponseException,
     PyiCloudServiceNotActivatedException,
 )
+from pyicloud_ipd.raw_policy import RawTreatmentPolicy
 from pyicloud_ipd.services.findmyiphone import AppleDevice, FindMyiPhoneServiceManager
 from pyicloud_ipd.services.calendar import CalendarService
 from pyicloud_ipd.services.ubiquity import UbiquityService
@@ -40,8 +41,6 @@ HEADER_DATA = {
     "scnt": "scnt",
 }
 
-
-
 class PyiCloudService:
     """
     A base authentication class for the iCloud service. Handles the
@@ -57,7 +56,7 @@ class PyiCloudService:
         self, 
         lp_filename_generator: Callable[[str], str],
         domain:str, 
-        use_raw_as_original: bool,
+        raw_policy: RawTreatmentPolicy,
         apple_id: str, password:Optional[str]=None, cookie_directory:Optional[str]=None, verify:bool=True,
         client_id:Optional[str]=None, with_family:bool=True,
     ):
@@ -65,7 +64,7 @@ class PyiCloudService:
             password = get_password_from_keyring(apple_id)
 
         self.lp_filename_generator = lp_filename_generator
-        self.use_raw_as_original = use_raw_as_original
+        self.raw_policy = raw_policy
         self.user: Dict[str, Any] = {"accountName": apple_id, "password": password}
         self.data: Dict[str, Any] = {} 
         self.params: Dict[str, Any] = {}
@@ -473,7 +472,7 @@ class PyiCloudService:
         """Gets the 'Photo' service."""
         if not self._photos:
             service_root = self._get_webservice_url("ckdatabasews")
-            self._photos = PhotosService(service_root, self.session, self.params, self.lp_filename_generator, self.use_raw_as_original)
+            self._photos = PhotosService(service_root, self.session, self.params, self.lp_filename_generator, self.raw_policy)
         return self._photos
 
     @property