+++ src/pyicloud_ipd/services/photos.py
@@ -17,6 +17,7 @@ import pytz
 
 from urllib.parse import urlencode
 
+from pyicloud_ipd.raw_policy import RawTreatmentPolicy
 from pyicloud_ipd.session import PyiCloudSession
 from pyicloud_ipd.utils import disambiguate_filenames
 
@@ -236,7 +237,7 @@ class PhotosService(PhotoLibrary):
 
     This also acts as a way to access the user's primary library.
     """
-    def __init__(self, service_root: str, session: PyiCloudSession, params: Dict[str, Any], lp_filename_generator: Callable[[str], str], use_raw_as_original: bool):
+    def __init__(self, service_root: str, session: PyiCloudSession, params: Dict[str, Any], lp_filename_generator: Callable[[str], str], raw_policy: RawTreatmentPolicy):
         self.session = session
         self.params = dict(params)
         self._service_root = service_root
@@ -246,7 +247,7 @@ class PhotosService(PhotoLibrary):
 
         self._libraries: Optional[Dict[str, PhotoLibrary]] = None
         self.lp_filename_generator = lp_filename_generator
-        self.use_raw_as_original = use_raw_as_original
+        self.raw_policy = raw_policy
 
         self.params.update({
             'remapEnums': True,
@@ -715,8 +716,8 @@ class PhotoAsset(object):
 
                     _versions[key] = version
 
-            # swap original & alternative if raw is alt & flag is set
-            if "alternative" in _versions and "raw" in _versions["alternative"]["type"] and self._service.use_raw_as_original:
+            # swap original & alternative according to swap_raw_policy
+            if "alternative" in _versions and (("raw" in _versions["alternative"]["type"] and self._service.raw_policy == RawTreatmentPolicy.AS_ORIGINAL) or ("raw" in _versions["original"]["type"] and self._service.raw_policy == RawTreatmentPolicy.AS_ALTERNATIVE)):
                 _a = dict(_versions["alternative"])
                 _o = dict(_versions["original"])
                 _versions["alternative"] = _o